Reorganizes Code and Adds Date Range Selection
This reorganizes code by splitting the app into the Component.js file
and manifest.json files. Also cleans up controller code for the
workitems controller. Adds a date range selection to the dialog
fragment which is presented to the user when they choose the filter /
sort button. This filter is then applied through table binding in
the workitems controller. Commit also adds an animated loading busy
indicator so that the user is aware that new data is being fetched
from the server.