Modifies Table View to List View
Changes the main work items display to show as a list. Combines
multiple attributes and sets two statuses for each item. One
represents the date and another status is shown in the form type. The
date is handled via several custom formatters to set colors and icons.
Updates the handleRow method in the workitems controller to correctly
launch the form.